Assignment 3/hardwarezone/requirements.txt
Scrapy==2.5.0
pymongo==3.12.0
kafka-python==2.0.2